L1: bugfixes in performance calculation

+++ reco/L1/CbmL1Constants.h
@@ -8,10 +8,10 @@
 namespace CbmL1Constants
 {
   /// Performance constants
-  const double MinRecoMom   = 0.1;  // Extra set of tracks = (MinRecoMom, MinRefMom)
-  const double MinFastMom   = 1.;   // Primary set of tracks = (MinRefMom, +inf)  //All reco tracks = (MinRecoMom, +inf)
-  const double MinPurity    = 0.7;  // min NStationInRecoTrack/NStationInMCTrack
-  const double MinNStations = 4.;   // min number of stations in track to be reconstructable
+  const double MinRecoMom = 0.1;  // Extra set of tracks = (MinRecoMom, MinRefMom)
+  const double MinFastMom = 1.;   // Primary set of tracks = (MinRefMom, +inf)  //All reco tracks = (MinRecoMom, +inf)
+  const double MinPurity  = 0.7;  // min NStationInRecoTrack/NStationInMCTrack
+  const int MinNStations  = 4;    // min number of stations on track to be reconstructable
 }  // namespace CbmL1Constants
 
 #endif
